Added Enemies and Player Shooting
We now have enemies that spawn and move about randomly with collision detection with the player. 
The player can now shoot at the enemies and destroy them with a magnificent animation. 
There is also a counter the keeps track of the number of enemies remaining. 
Added a features/todo list.