Updated Damage System and Coin Distribution
I’ve updated the fire rate to be slower for the initial character, but they can purchase damage power-ups from the store to reduce the fire rate. The coin distribution has been reworked to where you receive 1/4 of a coin per kill and 5 coins for killing a guardian. The guardians health is now double the health of the players and it’s damage is now triple.

Signed-off-by: Dalton Childers <dalton.childers@ssrc.msstate.edu>